+++ CHANGELOG.md
@@ -8,7 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fix jwt error returning HTTP status `400` for invalid role by @taimoorzaeem in #3601
-- Allow `db-extra-search-path` to accept empty value by @taimoorzaeem in #4074
+- Fix `db-extra-search-path` cannot be set to nothing by @taimoorzaeem in #4074
+  + It can now be disabled by setting it to empty string.
+  + Schema Cache load error is now logged including `db-schemas` and `db-extra-search-path` config values.
 
 ## [13.0.0] - 2025-05-08